+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Compatibility:
 * Decide the visibility in `Module#define_method` based on `self` and the default definee (#2334).
 * Configure `mandir` value in `RbConfig::CONFIG` and `RbConfig::MAKEFILE_CONFIG` (#2315).
 * TruffleRuby now supports the Truffle polyglot Hash interop API.
+* Implement `Fiber#raise` (#2338).
 
 Performance: